requirements.txt
feedparser==6.0.10
aiohttp==3.8.5
discord.py==2.3.1